refactor ui serializers
