+++ nanome/_internal/__init__.py
@@ -13,5 +13,4 @@ from .addon import _Addon  # noqa F401
 from .plugin_instance import _PluginInstance  # noqa F401
 from . import plugin_instance_deprecated  # noqa F401
 from .room import _Room  # noqa F401
-from .files import _Files  # noqa F401
 from .plugin import _Plugin  # noqa F401
