--- nanome/_internal/files.py
@@ -1,3 +1,0 @@
-class _Files():
-    def __init__(self):
-        pass
