+++ nanome/_internal/ui/__init__.py
@@ -17,5 +17,5 @@ from .menu import _Menu
 # folders
 from . import *
 from . import _io
-from . import _serialization
+from . import serializers
 from . import _deprecated
