--- nanome/_internal/ui/_serialization/dropdown_item_serializer.py
@@ -1,27 +1,0 @@
-from .. import _DropdownItem
-from nanome._internal.util.type_serializers import StringSerializer, TypeSerializer
-
-
-class _DropdownItemSerializer(TypeSerializer):
-    def __init__(self):
-        self.string = StringSerializer()
-
-    def version(self):
-        return 0
-
-    def name(self):
-        return "DropdownItem"
-
-    def serialize(self, version, value, context):
-        context.write_using_serializer(self.string, value._name)
-        context.write_bool(value._selected)
-        context.write_bool(value._close_on_selected)
-        context.write_byte(1)
-
-    def deserialize(self, version, context):
-        value = _DropdownItem._create()
-        value._name = context.read_using_serializer(self.string)
-        value._selected = context.read_bool()
-        value._close_on_selected = context.read_bool()
-        context.read_byte()  # eat the type for now, since it isn't supported quite yet
-        return value
